requirements.txt
opencv-python~=4.9.0.80
git+https://github.com/ageitgey/face_recognition_models
face-recognition
mysql-connector-python~=8.1.0
psycopg2
gTTS~=2.3.2
playsound~=1.3.0
xlsxwriter~=3.1.2
excel2json-3
pandas~=2.0.3
numpy~=1.26.4
pydub~=0.25.1
DeepImageSearch~=2.5
flask
pillow
watchdog~=4.0.0
pyttsx3~=2.90
googletrans==4.0.0-rc1
PyYAML~=6.0.1
